% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculate indices for a single game}
\usage{
calculate(data, game_id, prep_fun, ...)
}
\arguments{
\item{data}{The dataset used to calculate indices.}

\item{game_id}{The id of game to be processed.}

\item{prep_fun}{A name (or symbol) of function.}

\item{...}{Further input for the processing function.}
}
\value{
A \link[tibble:tibble-package]{tibble}. A new variable called "indices",
each of which contains a \link[tibble:tibble-package]{tibble} row of calculated
indices, is added to the data.
}
\description{
Given a game and a function name, \code{calculate()} extract data of that game and
calculate indices accordingly.
}
\author{
Liang Zhang (psychelzh@outlook.com)
}
